:sparkles: chain, non-chain에 맞게 패키지 이동및 구현
