+++ src/main/java/nextstep/auth/interceptor/chain/AuthChainInterceptor.java
@@ -1,0 +1,24 @@
+package nextstep.auth.interceptor.chain;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import nextstep.auth.context.Authentication;
+import nextstep.auth.context.SecurityContextHolder;
+import org.springframework.web.servlet.HandlerInterceptor;
+
+public abstract class AuthChainInterceptor implements HandlerInterceptor {
+
+    @Override
+    public boolean preHandle(final HttpServletRequest request, final HttpServletResponse response, final Object handler) throws Exception {
+        try {
+            final Authentication authentication = createAuthentication(request);
+            SecurityContextHolder.getContext().setAuthentication(authentication);
+            return true;
+        } catch (Exception e) {
+            return true;
+        }
+    }
+
+    protected abstract Authentication createAuthentication(HttpServletRequest request);
+
+}
